requirements.txt
streamlit
pandas
PyPDF2
google-generativeai
speech_recognition
streamlit-mic-recorder
openpyxl